+++ src/blob/handlers/BlockBlobHandler.ts
@@ -298,7 +298,14 @@ export default class BlockBlobHandler
     if (rawBody === undefined) {
       throw badRequestError;
     }
-    const parsed = await parseXML(rawBody, true);
+
+    let parsed;
+    try {
+      parsed = await parseXML(rawBody, true);
+    } catch (err) {
+      // return the 400(InvalidXmlDocument) error for issue 1955
+      throw StorageErrorFactory.getInvaidXmlDocument(context.contextId);
+    }
 
     // Validate selected block list
     const commitBlockList = [];