+++ ChangeLog.md
@@ -9,6 +9,8 @@ Blob:
 - Fix validation of Blob SAS token when using the second key for an account in `AZURITE_ACCOUNTS`
 - Set accessTierInferred to false after upload blob with accessTier (issue #2038)
 - Support blob new access tier Cold
+- Fixed startCopyFromURL, copyFromURL API to return 400 (InvalidHeaderValue) when copy source has invalid format. (issue #1954)
+- Fixed CommitBlockList API to return 400 (InvalidXmlDocument) when the request is sent with JSON body. (issue #1955)
 - Added "x-ms-is-hns-enabled" header in x-ms-is-hns-enabled API responds (issue #1810)
 
 Queue: